Create armstrong.py
a function (or method), that will take a number and returns true if that number is Armstrong number, or false
otherwise. An Armstrong number is a 3 digit number for which sum of cube of its digit is equal to the number. Example of
Armstrong number is 153 as   153= 1+ 125+27   which   1^3+5^3+3^3  . Another Armstrong number is 371.